Pseudopotentials.PSlibrary module

Pseudopotentials.PSlibrary module

Pseudopotentials.PSlibrary.download_potential — Method

download_potential(element::AbstractString, root::AbstractString)

Download one or multiple pseudopotentials from PSlibrary for a specific element under the same root.

Pseudopotentials.PSlibrary.download_potential — Method

download_potential(element::AbstractString)
download_potential(i::Integer)

Download one or multiple pseudopotentials from PSlibrary for a specific element.

Pseudopotentials.PSlibrary.list_elements — Method

list_elements()

List all elements that has pseudopotentials available in PSlibrary.

Pseudopotentials.PSlibrary.list_potential — Function

list_potential(element[, db])

List all pseudopotentials in PSlibrary for a specific element (abbreviation or index).

Arguments

element::Union{AbstractString,AbstractChar,Integer}: the element to find pseudopotentials with. The integer corresponding to the element's atomic index.

db::AbstractString="$element.jld2": the path to the database file.

See also: save_potential

Pseudopotentials.PSlibrary.save_potential — Function

save_potential(element, file[, db])

Save a PseudopotentialFile to the element's list.

Arguments

element::Union{AbstractString,Integer}: the element to save pseudopotentials with. The integer corresponding to the element's atomic index.

file::PseudopotentialFile: the object that stores the information of that file.

db::AbstractString="$element.jld2": the path to the database file.

See also: list_potential