Pseudopotentials module

Pseudopotentials module

Pseudopotentials.AndreaDalCorso — Type

AndreaDalCorso <: PseudopotentialFormat

A singleton representing the Andrea Dal Corso's code (old format).

Pseudopotentials.OldNormConserving — Type

OldNormConserving <: PseudopotentialFormat

A singleton representing the old PWscf norm-conserving format.

Pseudopotentials.UnifiedPseudopotentialFormat — Type

UnifiedPseudopotentialFormat <: PseudopotentialFormat

A singleton representing the new UPF format. If it doesn't work, the pseudopotential format is determined by the file name.

Pseudopotentials.VanderbiltUltraSoft — Type

VanderbiltUltraSoft <: PseudopotentialFormat

A singleton representing the Vanderbilt US pseudopotential code.

Pseudopotentials.pseudopot_format — Method

pseudopot_format(data::AbstractString)

Return the pseudopotential format.

The pseudopotential file is assumed to be in the new UPF format. If it doesn't work, the pseudopotential format is determined by the file name:

"*.vdb or *.van": Vanderbilt US pseudopotential code

"*.RRKJ3": Andrea Dal Corso's code (old format)

none of the above: old PWscf norm-conserving format

Pseudopotentials.PseudopotentialFormat — Type

PseudopotentialFormat

Represent all possible pseudopotential file formats.